Html mods (#2)
* Updated summary about website and added pictures

new summary was about the website was done to allow users to understand what the system is capable of doing.
new pictures were added to make it visually appealing.
a background image could be added down the line as well as a modified 'contact us' section.

* updated 'contact us' details

updated the 'contact us' details such as the address and email address.
the working hours were also updated as well.

* Added form and other items to file

this file is to query the database for current inventory amounts for AN item.
drop down box was used to select the items to query.

future additions that could be done are auto update inventory items

* new update_inventory file to update inventory amounts

with this file, the users can now update the amount of inventory for aech inventory items.
future changes include the adding of php to pull values from database to display the current value of the item
as well as to update the database

* added html page for adding new inventory

users can now create new inventory items. future changes to make include updating
delete_inv view_current_inv and update_inv to reflect new items made or deleted.

* added html page for user to delete inventory item

with this form, the user can completely delete an item from the database instead of
setting the amount of individual item(s) to 0.

* changed heading of the page

fixed a mistake mis-naming the page

* updated Index with pictures and links to pages

there are now links to the different pages for the working off the management system. Pictures
were added as well for the fun of it. Unrelated data and pictures were deleted.

* Added name and value to button

Change was done to make it easier to call php script when button is pressed.
php change coming up will allow program to call database to print out current inventory

* prepped files to add php code/files

unrequired lines of code were deleted to add new php code down the line.
comments were added describing the requirements of the php code.